skill/pyproject.toml
[project]
name = "text2sql-skill"
license = "Apache-2.0"
version = "0.1.0"
authors = [{ name = "Aleph Alpha Engineering", email = "engineering@aleph-alpha.com" }]
description = "Skill Example"

requires-python = "~=3.11"

dependencies = [
    "python-liquid==1.9.4",
    "colorama>=0.4.6",
    "pharia-skill>=0.20.0",
    "jinja2>=3.1.6",
    "black>=25.1.0",
    "matplotlib>=3.7.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pharia-inference-sdk>=0.1.6",
    "pharia-studio-sdk>=0.1.7",
]

[dependency-groups]
dev = [
    "pre-commit==4.0.1",
    "ruff==0.9.4",
    "mypy==1.11.2",
    "pytest==8.3.3",
    "python-dotenv==1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.hatch.build.targets.sdist]
include = ["."]
